# application/services/custom_services/gporder/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: gepordersnew open api specification
   description: gporder API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8014
     description: gporder
tags:
- name: gporders
  description: order items
- name: order
  description: represents and order
paths:
   "/gporders/{id}":
      delete:
         tags:
         -  gporders
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  gporders
         description: GpGetNounById
         parameters:
         -  in: "path"
            description: GpGetNounById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/gporders/get/search":
      get:
         tags:
         -  gporders
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/gporders/get/update":
      put:
         tags:
         -  gporders
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/gporders"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/gporders":
      put:
         tags:
         -  gporders
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/gporders"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  gporders
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  gporders
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/gporders"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/gporders/userid/created_by":
      get:
         tags:
         -  gporders
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      gporders:
         type: object
         properties:
            type:
               type: "string"
               description: Description
            name:
               type: "string"
               description: Description
            description:
               type: "string"
               description: Description
            price:
               type: "number"
               description: Description
            price_currency_type:
               type: "string"
               description: Description
            cost:
               type: "number"
               description: Description
            cost_currency_type:
               type: "string"
               description: Description
            gephistoryid:
               type: "string"
               description: Description
      order:
         type: object
         properties:
            type:
               type: "string"
               description: Description
            orderstatus:
               type: "string"
               description: Description
            name:
               type: "string"
               description: Description
            description:
               type: "string"
               description: Description
            attachments:
               type: "string"
               description: Description
            gephistoryid:
               type: "string"
               description: Description